refactor(tests): organize tests by type
subdirectories contain different types of tests:

* integration - with external systems
* ui - the web app and its pages
