+++ tests/e2e/cypress/specs/ui/help.spec.js
@@ -1,4 +1,4 @@
-const agencies = require("../fixtures/transit-agencies");
+const agencies = require("../../fixtures/transit-agencies");
 
 describe("Help page spec", () => {
   beforeEach(() => {
