+++ tests/e2e/cypress/specs/ui/payment-options.spec.js
